+++ Navigator.js
@@ -1,0 +1,11 @@
+import {
+    createAppContainer,
+    createSwitchNavigator,
+} from 'react-navigation';
+import Home from './screens';
+
+const Navigator = createSwitchNavigator({
+    Home,
+});
+
+export default createAppContainer(Navigator);
