+++ README.md
@@ -1,4 +1,6 @@
 PoddyCam
-########
+========
 
 Just a way to spook our puppers when we can't.
+
+![cute-beagle](https://media.giphy.com/media/8CVM3XzTzptjG/giphy.gif)
