+++ package.json
@@ -11,6 +11,7 @@
     "prop-types": "^15.6.2",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz",
+    "react-navigation": "^3.0.5",
     "styled-components": "^4.1.2"
   },
   "devDependencies": {
