+++ styles/palette.js
